+++ packages/edit-site/src/components/save-button/index.js
@@ -7,7 +7,10 @@ import { __, _n, sprintf } from '@wordpress/i18n';
 import { store as coreStore } from '@wordpress/core-data';
 import { displayShortcut } from '@wordpress/keycodes';
 import { privateApis as routerPrivateApis } from '@wordpress/router';
-import { useEntitiesSavedStatesIsDirty } from '@wordpress/editor';
+import {
+	useEntitiesSavedStatesIsDirty,
+	store as editorStore,
+} from '@wordpress/editor';
 
 /**
  * Internal dependencies
@@ -32,7 +35,7 @@ export default function SaveButton( {
 } ) {
 	const { params } = useLocation();
 	const { setIsSaveViewOpened } = useDispatch( editSiteStore );
-	const { saveDirtyEntities } = unlock( useDispatch( coreStore ) );
+	const { saveDirtyEntities } = unlock( useDispatch( editorStore ) );
 	const { dirtyEntityRecords } = useEntitiesSavedStatesIsDirty();
 	const { isSaving, isSaveViewOpen, previewingThemeName } = useSelect(
 		( select ) => {